+++ Classes/CrawlStrategy/GuzzleExecutionStrategy.php
@@ -31,7 +31,7 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
 /**
  * Calls Guzzle / CURL (based on TYPO3 settings) for fetching a URL.
  */
-class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategy
+class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategyInterface
 {
     use LoggerAwareTrait;